Add request caching to Dockerfile generator
When developing parts of the Dockerfile generation pipeline, and thus regenerating Dockerfiles over and over, many
unnecessary HTTP requests are being made to https://api.adoptium.net, since the data is mostly unchanged. Adding a
request cache with an expiration time of an hour allows this regeneration to run in under a second in most case, while
not compromising on the correctness.