Fix CA certificate processing when a certificate is a symlink
Usually, a certificate in a mount will be a regular file. However in some environments, like when providing certificates
via a `ConfigMap` in a Kubernetes context, the certificate file will be a symlink. Our CA certificate processing copies
certificate files to a new location and since the symlink targets are hidden files in the same directory, those do not
get copied along and symlinks could become broken. This patch dereferences symlinks when copying to avoid broken links.

Signed-off-by: Nikolai Prokoschenko <nikolai.prokoschenko@kurzdigital.com>